Add files via upload
Our goal in this project is to build a binary classification based Logistic Regression model to automatically to predict whether a bank note is Genuine or Counterfeit. In this project we are using the Banknote Authentication Data Set provided by UCI Machine Learning Repository. Further in this project we will optimise the parameters of the Logistic Regression model by minimising the SoftMax cost function.